# perf/duplicate_probe: Run test against upstream kernel