--- a/perf/perf_duplicate_probe.py.data/perf_duplicate.yaml
+++ b/perf/perf_duplicate_probe.py.data/perf_duplicate.yaml
@@ -1,0 +1,5 @@
+run_type: !mux
+    upstream:
+        type: 'upstream'
+    distro:
+        type: 'distro'