This 2 patch series adds support to run the test against
upstream kernel. Second patch is a test case fixup.

perf/duplicate_probe: Run test against upstream kernel
perf/duplicate_probe: Delete probes only if previously installed

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>