@sacsant 2023-01-09
Before the fix:

 (1/1) perf/perf_duplicate_probe.py:PerfProbe.test_probe: STARTED
 (1/1) perf/perf_duplicate_probe.py:PerfProbe.test_probe: ERROR: Command 'perf probe -d \*' failed.\nstdout: b''\nstderr: b'"*" does not hit any event.\n  Error: Failed to delete events.\n'\nadditional_info: None (1.25 s)

After the fix:

 (1/1) perf/perf_duplicate_probe.py:PerfProbe.test_probe: STARTED
 (1/1) perf/perf_duplicate_probe.py:PerfProbe.test_probe: CANCEL: kernel-debuginfo is needed for the test to be run (1.25 s)

With distro as a target:

 (1/1) perf/perf_duplicate_probe.py:PerfProbe.test_probe: STARTED
 (1/1) perf/perf_duplicate_probe.py:PerfProbe.test_probe: PASS (0.95 s)

With upstream as a target:

 (1/1) perf/perf_duplicate_probe.py:PerfProbe.test_probe;run-run_type-upstream-2f21: STARTED
 (1/1) perf/perf_duplicate_probe.py:PerfProbe.test_probe;run-run_type-upstream-2f21: PASS (0.77 s)